+++ client/src/ts/modules/hangout/settings/hangoutSettings.ts
@@ -2,11 +2,15 @@ import axios, { AxiosError } from "../../../../../node_modules/axios/index";
 import { handleAuthSessionDestroyed, handleAuthSessionExpired } from "../../global/authUtils";
 import { dayMilliseconds, HANGOUT_AVAILABILITY_STAGE, HANGOUT_CONCLUSION_STAGE, HANGOUT_SUGGESTIONS_STAGE, HANGOUT_VOTING_STAGE, MAX_HANGOUT_MEMBERS_LIMIT, MAX_HANGOUT_PERIOD_DAYS, MIN_HANGOUT_MEMBERS_LIMIT } from "../../global/clientConstants";
 import { ConfirmModal } from "../../global/ConfirmModal";
+import ErrorSpan from "../../global/ErrorSpan";
+import { InfoModal } from "../../global/InfoModal";
 import LoadingModal from "../../global/LoadingModal";
 import popup from "../../global/popup";
 import SliderInput from "../../global/SliderInput";
-import { ProgressHangoutStageData, progressHangoutStageService, UpdateHangoutStagesBody, updateHangoutStagesService } from "../../services/hangoutServices";
+import { validateNewPassword, validatePassword } from "../../global/validation";
+import { ProgressHangoutStageData, progressHangoutStageService, updateHangoutMembersLimitService, updateHangoutPasswordService, UpdateHangoutStagesBody, updateHangoutStagesService } from "../../services/hangoutServices";
 import { globalHangoutState } from "../globalHangoutState";
+import { copyToClipboard } from "../globalHangoutUtils";
 import { directlyNavigateHangoutSections, navigateHangoutSections } from "../hangoutNav";
 import { hangoutSuggestionState } from "../suggestions/hangoutSuggestions";
 import { calculateStepMinimumSliderValue, handleProgressionAttemptWithoutSuggestions, isValidNewHangoutPeriods, resetMembersLimitSliderValues, resetSliderValues, resetStageSliderValues, toggleStagesSettingsButtons, updateSettingsButtons } from "./hangoutSettingsUtils";
@@ -39,8 +43,15 @@ export const hangoutSettingsState: HangoutSettingsState = {
 };
 
 const settingsSectionElement: HTMLElement | null = document.querySelector('#settings-section');
+
 const updateHangoutPasswordForm: HTMLFormElement | null = document.querySelector('#hangout-settings-password-form');
+const settingsPasswordInput: HTMLInputElement | null = document.querySelector('#hangout-settings-password-input');
+
 const progressHangoutBtn: HTMLButtonElement | null = document.querySelector('#progress-hangout-btn');
+const membersCountSpan: HTMLSpanElement | null = document.querySelector('#settings-member-count-span');
+
+const settingsPasswordPreviewer: HTMLSpanElement | null = document.querySelector('#settings-password-previewer');
+const deleteHangoutPassword: HTMLButtonElement | null = document.querySelector('#delete-hangout-password-btn');
 
 export function hangoutSettings(): void {
   loadEventListeners();
@@ -50,7 +61,10 @@ function loadEventListeners(): void {
   document.addEventListener('loadSection-settings', initHangoutSettings);
 
   settingsSectionElement?.addEventListener('click', handleHangoutSettingsClicks);
-  updateHangoutPasswordForm?.addEventListener('submit', updateHangoutPassword);
+  updateHangoutPasswordForm?.addEventListener('submit', async (e: SubmitEvent) => {
+    e.preventDefault();
+    updateHangoutPassword();
+  });
 };
 
 function initHangoutSettings(): void {
@@ -122,6 +136,8 @@ function initHangoutSettings(): void {
   document.addEventListener('members-limit-input_valueChange', updateSettingsButtons);
 
   renderHangoutSettingsSection();
+  setActiveValidation();
+
   LoadingModal.remove();
 };
 
@@ -129,6 +145,8 @@ function renderHangoutSettingsSection(): void {
   updateSliderValues();
   disablePassedStagesSliders();
   updateProgressBtn();
+  updateMembersCount();
+  updateHangoutPasswordElements();
 
   if (!hangoutSettingsState.settingsSectionMutationObserverActive) {
     initSettingsSectionMutationObserver();
@@ -179,6 +197,38 @@ function updateProgressBtn(): void {
   };
 };
 
+function updateMembersCount(): void {
+  if (!hangoutSettingsState.sliders || !membersCountSpan) {
+    return;
+  };
+
+  membersCountSpan.textContent = `${hangoutSettingsState.sliders.membersLimitSlider.value}`;
+};
+
+function updateHangoutPasswordElements(): void {
+  if (!globalHangoutState.data || !settingsPasswordPreviewer || !deleteHangoutPassword) {
+    return;
+  };
+
+  const decryptedHangoutPassword: string | null = globalHangoutState.data.decryptedHangoutPassword;
+
+  if (!decryptedHangoutPassword) {
+    settingsPasswordPreviewer.classList.add('empty');
+    settingsPasswordPreviewer.firstElementChild && (settingsPasswordPreviewer.firstElementChild.textContent = 'Not set');
+
+    deleteHangoutPassword.classList.add('disabled');
+    deleteHangoutPassword.disabled = true;
+
+    return;
+  };
+
+  settingsPasswordPreviewer.classList.remove('empty');
+  settingsPasswordPreviewer.firstElementChild && (settingsPasswordPreviewer.firstElementChild.textContent = '*************');
+
+  deleteHangoutPassword.classList.remove('disabled');
+  deleteHangoutPassword.disabled = false;
+};
+
 async function handleHangoutSettingsClicks(e: MouseEvent): Promise<void> {
   if (!(e.target instanceof HTMLButtonElement)) {
     return;
@@ -215,17 +265,22 @@ async function handleHangoutSettingsClicks(e: MouseEvent): Promise<void> {
   };
 
   if (e.target.id === 'settings-password-copy-btn') {
-    // TODO: implement
+    await handleCopyHangoutPassword();
+    return;
+  };
+
+  if (e.target.id === 'settings-password-reveal-btn') {
+    revealHangoutPassword();
     return;
   };
 
   if (e.target.id === 'hangout-settings-password-input-reveal-btn') {
-    // TODO: implement
+    togglePasswordInputReveal(e.target);
     return;
   };
 
   if (e.target.id === 'delete-hangout-password-btn') {
-    // TODO: implement
+    confirmPasswordDelete();
   };
 };
 
@@ -477,11 +532,226 @@ async function progressHangoutStage(): Promise<void> {
 };
 
 async function updateHangoutMembersLimit(): Promise<void> {
-  // TODO: implement
+  LoadingModal.display();
+
+  if (!globalHangoutState.data || !hangoutSettingsState.sliders) {
+    popup('Something went wrong.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  const { hangoutId, hangoutMemberId, hangoutDetails, hangoutMembers } = globalHangoutState.data;
+
+  const currentMembersCount: number = hangoutMembers.length;
+  const currentMembersLimit: number = hangoutDetails.members_limit;
+  const newMembersLimit: number = hangoutSettingsState.sliders.membersLimitSlider.value;
+
+  if (hangoutDetails.is_concluded) {
+    popup('Hangout has already been concluded.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  if (newMembersLimit === currentMembersLimit) {
+    updateSettingsButtons();
+
+    popup(`Members limit is already set to ${currentMembersLimit}.`, 'success');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  if (newMembersLimit < currentMembersCount) {
+    popup(`New members limit can't be lower than the number of existing members.`, 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  try {
+    await updateHangoutMembersLimitService({ hangoutId, hangoutMemberId, newMembersLimit });
+
+    hangoutDetails.members_limit = newMembersLimit;
+    hangoutSettingsState.sliders.membersLimitSlider.updateValue(newMembersLimit);
+
+    updateMembersCount();
+
+    popup('Members limit updated.', 'success');
+    LoadingModal.remove();
+
+  } catch (err: unknown) {
+    console.log(err);
+    LoadingModal.remove();
+
+    if (!axios.isAxiosError(err)) {
+      popup('Something went wrong.', 'error');
+      return;
+    };
+
+    const axiosError: AxiosError<AxiosErrorResponseData> = err;
+
+    if (!axiosError.status || !axiosError.response) {
+      popup('Something went wrong.', 'error');
+      return;
+    };
+
+    const status: number = axiosError.status;
+    const errMessage: string = axiosError.response.data.message;
+    const errReason: string | undefined = axiosError.response.data.reason;
+
+    if (status === 400) {
+      popup('Something went wrong/.', 'error');
+      return;
+    };
+
+    popup(errMessage, 'error');
+
+    if (status === 403) {
+      globalHangoutState.data.hangoutDetails.current_stage = HANGOUT_CONCLUSION_STAGE;
+      globalHangoutState.data.hangoutDetails.is_concluded = true;
+
+      renderHangoutSettingsSection();
+      return;
+    };
+
+    if (status === 404) {
+      LoadingModal.display();
+      setTimeout(() => window.location.reload(), 1000);
+
+      return;
+    };
+
+    if (status === 401) {
+      if (errReason === 'notHangoutLeader') {
+        globalHangoutState.data.isLeader = false;
+        directlyNavigateHangoutSections('dashboard');
+
+        return;
+      };
+
+      if (errReason === 'authSessionExpired') {
+        handleAuthSessionExpired();
+        return;
+      };
+
+      handleAuthSessionDestroyed();
+    };
+  };
 };
 
-async function updateHangoutPassword(): Promise<void> {
-  // TODO: implement
+async function updateHangoutPassword(deletePassword: boolean = false): Promise<void> {
+  LoadingModal.display();
+
+  if (!globalHangoutState.data || !settingsPasswordInput) {
+    popup('Something went wrong.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  const { hangoutId, hangoutMemberId, isLeader, hangoutDetails } = globalHangoutState.data;
+
+  if (!isLeader) {
+    popup(`You're not the hangout leader.`, 'error');
+    LoadingModal.remove();
+
+    directlyNavigateHangoutSections('dashboard');
+    return;
+  };
+
+  if (hangoutDetails.is_concluded) {
+    popup('Hangout has already been concluded.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  const isValidNewPassword: boolean = deletePassword ? true : validatePassword(settingsPasswordInput);
+
+  if (!isValidNewPassword) {
+    popup('Invalid new hangout password.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  const newPassword: string | null = deletePassword ? null : settingsPasswordInput.value;
+
+  try {
+    await updateHangoutPasswordService({ hangoutId, hangoutMemberId, newPassword });
+    globalHangoutState.data.decryptedHangoutPassword = newPassword;
+
+    clearUpdatePasswordForm();
+    updateHangoutPasswordElements();
+
+    popup(deletePassword ? 'Password removed.' : 'Password updated.', 'success');
+    LoadingModal.remove();
+
+  } catch (err: unknown) {
+    console.log(err);
+    LoadingModal.remove();
+
+    if (!axios.isAxiosError(err)) {
+      popup('Something went wrong.', 'error');
+      return;
+    };
+
+    const axiosError: AxiosError<AxiosErrorResponseData> = err;
+
+    if (!axiosError.status || !axiosError.response) {
+      popup('Something went wrong.', 'error');
+      return;
+    };
+
+    const status: number = axiosError.status;
+    const errMessage: string = axiosError.response.data.message;
+    const errReason: string | undefined = axiosError.response.data.reason;
+
+    if (status === 400 && !errReason) {
+      popup('Something went wrong.', 'error');
+      return;
+    };
+
+    popup(errMessage, 'error');
+
+    if (status === 403) {
+      globalHangoutState.data.hangoutDetails.current_stage = HANGOUT_CONCLUSION_STAGE;
+      globalHangoutState.data.hangoutDetails.is_concluded = true;
+
+      renderHangoutSettingsSection();
+      return;
+    };
+
+    if (status === 400) {
+      ErrorSpan.display(settingsPasswordInput, errMessage);
+      return;
+    };
+
+    if (status === 404) {
+      LoadingModal.display();
+      setTimeout(() => window.location.reload(), 1000);
+
+      return;
+    };
+
+    if (status === 401) {
+      if (errReason === 'notHangoutLeader') {
+        globalHangoutState.data.isLeader = false;
+        directlyNavigateHangoutSections('dashboard');
+
+        return;
+      };
+
+      if (errReason === 'authSessionExpired') {
+        handleAuthSessionExpired();
+        return;
+      };
+
+      handleAuthSessionDestroyed();
+    };
+  };
 };
 
 function initSettingsSectionMutationObserver(): void {
@@ -549,4 +819,116 @@ function confirmProgressHangoutAction(): void {
       ConfirmModal.remove();
     };
   });
+};
+
+async function handleCopyHangoutPassword(): Promise<void> {
+  if (!globalHangoutState.data) {
+    popup('Failed to copy hangout password.', 'error');
+    return;
+  };
+
+  const { isLeader, decryptedHangoutPassword } = globalHangoutState.data;
+
+  if (!isLeader) {
+    popup(`You're not the hangout leader.`, 'error');
+    directlyNavigateHangoutSections('dashboard');
+
+    return;
+  };
+
+  if (!decryptedHangoutPassword) {
+    popup('Hangout is not password protected.', 'error');
+    updateHangoutPasswordElements();
+
+    return;
+  };
+
+  await copyToClipboard(decryptedHangoutPassword);
+};
+
+function revealHangoutPassword(): void {
+  if (!globalHangoutState.data) {
+    popup('Failed to reveal hangout password.', 'error');
+    return;
+  };
+
+  const { isLeader, decryptedHangoutPassword } = globalHangoutState.data;
+
+  if (!isLeader) {
+    popup(`You're not the hangout leader.`, 'error');
+    directlyNavigateHangoutSections('dashboard');
+
+    return;
+  };
+
+  if (!decryptedHangoutPassword) {
+    popup('Hangout is not password protected.', 'error');
+    updateHangoutPasswordElements();
+
+    return;
+  };
+
+  InfoModal.display({
+    title: 'Hangout password is: ',
+    description: decryptedHangoutPassword,
+    btnTitle: 'Hide',
+  }, { simple: true });
+};
+
+function confirmPasswordDelete(): void {
+  const confirmModal: HTMLDivElement = ConfirmModal.display({
+    title: 'Are you sure you want to remove the hangout password?',
+    description: 'This action will make your hangout less secure.',
+    confirmBtnTitle: 'Remove password',
+    cancelBtnTitle: 'Cancel',
+    extraBtnTitle: null,
+    isDangerousAction: true,
+  });
+
+  confirmModal.addEventListener('click', async (e: MouseEvent) => {
+    if (!(e.target instanceof HTMLButtonElement)) {
+      return;
+    };
+
+    if (e.target.id === 'confirm-modal-confirm-btn') {
+      await updateHangoutPassword(true);
+      ConfirmModal.remove();
+
+      return;
+    };
+
+    if (e.target.id === 'confirm-modal-cancel-btn') {
+      ConfirmModal.remove();
+    };
+  });
+};
+
+function setActiveValidation(): void {
+  settingsPasswordInput?.addEventListener('keyup', () => validateNewPassword(settingsPasswordInput));
+};
+
+function clearUpdatePasswordForm(): void {
+  if (!settingsPasswordInput) {
+    return;
+  };
+
+  settingsPasswordInput.blur();
+  ErrorSpan.hide(settingsPasswordInput);
+  settingsPasswordInput.value = '';
+};
+
+function togglePasswordInputReveal(revealBtn: HTMLButtonElement): void {
+  if (!revealBtn.previousElementSibling) {
+    return;
+  };
+
+  if (revealBtn.classList.contains('revealed')) {
+    revealBtn.classList.remove('revealed');
+    revealBtn.previousElementSibling.setAttribute('type', 'password');
+
+    return;
+  };
+
+  revealBtn.classList.add('revealed');
+  revealBtn.previousElementSibling.setAttribute('type', 'text');
 };